Some fixes that may address zombie pods
The watch loop kept starting with resource_version="" at every timeout
(2 minutes by default), this could be a possible cause of missed events
causing zombie pods. We now continue to track the resource_version when
an event is received so we resume from there.

The startup logic was also not processing any current events, it only
processed zombie or stale pods or backfilling. It now also starts
tracking running pods before handing over to the watch loop so there are
no gaps.

Additionally, if an exception is raised then we relaunch the startup
logic instead of just continuing the fresh loop so we can immediately
mark any zombie pods.

Another check is added so we don't start billing for pods that are set
to be terminated.